requirements.txt
git+https://github.com/psycopg/psycopg.git#subdirectory=psycopg
backports.zoneinfo==0.2.1
boto3==1.18.29
botocore==1.21.29
cffi==1.14.6
importlib-resources==5.2.2
jmespath==0.10.0
mysql-connector-python==8.0.26
pkg-resources==0.0.0
protobuf==3.17.3
pycparser==2.20
python-dateutil==2.8.2
s3transfer==0.5.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
